refactor: eliminar componentes de layout y categorías, actualizar importaciones
