+++ next.config.mjs
@@ -6,7 +6,7 @@ const withNextIntl = createNextIntlPlugin();
 
 const nextConfig = {
   images: {
-    domains: ["pub-843b15cc9c5d4932b34855ee68a2f5be.r2.dev"],
+    domains: ["pub-b968cd75d23f41bf90ea2c70ff12436f.r2.dev"],
   },
 };
