--- src/app/(shop)/layout.tsx
@@ -1,6 +1,0 @@
-import { LayoutShop as LayoutOnlyShop } from "@/components/templates/LayoutShop";
-export default async function LayoutShop({
-  children,
-}: Readonly<{ children: React.ReactNode }>) {
-  return <LayoutOnlyShop>{children}</LayoutOnlyShop>;
-}
