+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import LayoutShop from "./(shop)/layout";
+import { LayoutShop } from "@/components/templates/LayoutShop";
 import { ProductsFilters } from "@/components/categories/ProductsFilters";
 
 export default function Home() {
